Split 5 oversized god-files into focused single-responsibility units
Refactor the largest files in the codebase to comply with the ~500-line
guideline. No logic changes — only code movement to new translation units.

- SparkEngine.cpp (2,454 → 1,677): Extract gameplay system lifecycle
  (init/update/shutdown for 40+ subsystems) to GameplaySystemLifecycle.cpp
- PhysicsSystemQueries.cpp (1,612 → 717): Extract constraints, spatial
  queries, and motor control to 3 new files
- SceneSerializer.cpp (1,533 → 232): Extract binary and JSON serialization
  to BinarySceneSerializer.cpp and JSONSceneSerializer.cpp
- AnimationSystem.cpp (1,409 → 650): Extract state machine, IK solver,
  and skeletal animation to 3 new files
- CrashHandler.cpp (1,417 → 1,088): Extract GitHub upload code to
  CrashReportUploader.cpp
