+++ SparkEngine/Source/Physics/PhysicsSystem.cpp
@@ -5,7 +5,10 @@
  * @author Spark Engine Team
  * @date 2025
  *
- * Body/constraint management and spatial queries live in PhysicsSystemQueries.cpp.
+ * Body lifecycle, debug rendering, materials, serialization live in PhysicsSystemQueries.cpp.
+ * Constraint creation lives in PhysicsConstraints.cpp.
+ * Constraint motor control lives in PhysicsMotorControl.cpp.
+ * Raycasting, overlap, and sweep queries live in PhysicsSpatialQueries.cpp.
  * Shape creation and Jolt conversion helpers live in PhysicsShapeFactory.cpp.
  * PhysicsBody and PhysicsConstraint methods live in PhysicsBodyImpl.cpp.
  */
