+++ wiki/Home.md
@@ -123,12 +123,12 @@ SparkEngine is licensed under the [MIT License](https://github.com/Krilliac/Spar
 <!-- AUTO:stats -->
 | Metric | Count |
 |--------|-------|
-| Header files | 527 |
+| Header files | 529 |
 | ECS Components | 79 |
 | ECS Systems | 64 |
 | Editor Panels | 51 |
 | Test files | 209 |
 | Test cases | 2508+ |
 | Wiki pages | 66 |
-| *Last synced* | *2026-03-29 14:07* |
+| *Last synced* | *2026-03-29 15:20* |
 <!-- /AUTO:stats -->
